app.service 파일 분리: 도메인 서비스 사용 여부
